backend: moved RangeDto in api/schemas
